replaced logging

+++ PosterBot.js
@@ -1,12 +1,14 @@
 const { MongoDB, MongoClient } = require('mongodb');
 const { Telegraf } = require('telegraf');
 const { CronJob } = require('cron');
+const Logger = require('./logging');
 
 //Database stuff
 var url = "mongodb://localhost:27017/";
 
 //Dont share this key
 const key = require("./.secret/key.json");
+const logging = require('./logging');
 const bot = new Telegraf(key.key);
 
 //try /start in the bot's dms
@@ -21,7 +23,7 @@ bot.command('upload', (ctx) => {
     //just incase anyone but me tries anything wack
     if (ctx.from.id != key.User_ID) {
         ctx.reply("I'm sorry John");
-        console.log("Unauthorized use of bot");
+        Logger.warn("Unauthorized use of bot");
         return;
     }
 
@@ -33,7 +35,7 @@ bot.command('upload', (ctx) => {
 
     client.connect((err, db) => {
         if (err) throw err;
-        var dbo = db.db("Telegram_Test_DB");
+        var dbo = db.db("Telegram_DB");
         var pisspic = { url: link };
         dbo.collection("Piss").insertOne(pisspic);
         db.close();
@@ -41,7 +43,7 @@ bot.command('upload', (ctx) => {
     
     //confirming database recieved
     ctx.reply("Received!");
-    console.log('New record added');
+    Logger.log('New record added');
 
 });
 
@@ -52,32 +54,40 @@ var job = new CronJob('*/1 * * * *', function() {
     var client = new MongoClient(url);
 
     //basic connection
-    client.connect((err, db) => {
+    client.connect(async (err, db) => {
         if (err) throw err;
         var dbo = db.db("Telegram_DB");
 
         //finding records to post
-        var topost = dbo.collection("Piss").find().sort({_id:1}).limit(2);
+        var topost = await dbo.collection("Piss").find().sort({_id:1}).limit(2);
 
-        topost.forEach((o,e) => {
-            console.log("Link posted - " + o.url);
-            bot.telegram.sendMessage(key.Chat_ID, o.url);
-        });
+        var recordcount = await topost.count();
+        
+        if(recordcount > 0) {
 
-        dbo.collection("Piss").deleteOne({});
-        dbo.collection("Piss").deleteOne({});
-        console.log("records deleted");
+            //send messages
+            topost.forEach((o,e) => {
+                Logger.log("Link posted - " + o.url);
+                bot.telegram.sendMessage(key.Chat_ID, o.url);
+            });
 
+            //deleting records after posting
+            await dbo.collection("Piss").deleteOne({});
+            await dbo.collection("Piss").deleteOne({});
+            Logger.log("records deleted");
+        } else {
+            Logger.warn("DB empty, not posting")
+        }
 
-        //deleted records
-        db.close();
+        
+        await db.close();
     });
 }, null, true, 'Australia/Sydney');
 
 
 //Start the bot with a little message
 bot.launch().then((o, e) => {
-    console.log("Starting bot...");
+    Logger.log("Starting bot...");
     job.start();
 });
